requirements_base.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg[binary]==3.1.18
alembic==1.13.1
pydantic-settings==2.2.1
aio-pika==9.4.1
httpx==0.27.0
pytest==8.1.1
pytest-cov==5.0.0
